DigiApp_3Page_StarterPack/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
pandas==2.1.4
python-dotenv==1.0.0
sendgrid==6.10.0
openai==1.3.0
requests==2.31.0
gunicorn==21.2.0
twilio==8.10.0 